Separated result rendering from controller actions.
