+++ cmd/csaf_provider/tmpl/create.html
@@ -8,7 +8,18 @@
   <body>
     <h1>CSAF-Provider - Directory structure created</h1>
     {{ if .Error }}
-    <strong>Error: <tt>{{ .Error }}.</tt></strong>
+    {{ if eq (len .Error) 1 }}
+    <strong>Error: <tt>{{ index .Error 0 }}.</tt></strong>
+    {{ else }}
+    <p>
+    Errors:
+    <ul>
+    {{ range .Error }}
+    <li>{{ . }}</li>
+    {{ end }}
+    </ul>
+    <p>
+    {{ end }}
     {{ else }}
     Everything is setup fine now.
     {{ end }}
